third_party/WebKit/LayoutTests/imported/wpt/webstorage/idlharness-expected.txt

This is a testharness.js-based test.
FAIL Window interface: attribute sessionStorage assert_equals: "sessionStorage" must have a getter expected "function" but got "undefined"
FAIL Window interface: attribute localStorage assert_equals: "localStorage" must have a getter expected "function" but got "undefined"
PASS Storage interface: existence and properties of interface object 
PASS Storage interface object length 
PASS Storage interface object name 
FAIL Storage interface: existence and properties of interface prototype object assert_equals: class string of Storage.prototype expected "[object StoragePrototype]" but got "[object Storage]"
PASS Storage interface: existence and properties of interface prototype object's "constructor" property 
FAIL Storage interface: attribute length assert_true: property is not enumerable expected true got false
FAIL Storage interface: operation key(unsigned long) assert_true: property is not enumerable expected true got false
FAIL Storage interface: operation getItem(DOMString) assert_true: property is not enumerable expected true got false
FAIL Storage interface: operation setItem(DOMString,DOMString) assert_true: property is not enumerable expected true got false
FAIL Storage interface: operation removeItem(DOMString) assert_true: property is not enumerable expected true got false
FAIL Storage interface: operation clear() assert_true: property is not enumerable expected true got false
PASS Storage must be primary interface of window.localStorage 
PASS Stringification of window.localStorage 
PASS Storage interface: window.localStorage must inherit property "length" with the proper type (0) 
PASS Storage interface: window.localStorage must inherit property "key" with the proper type (1) 
PASS Storage interface: calling key(unsigned long) on window.localStorage with too few arguments must throw TypeError 
PASS Storage interface: window.localStorage must inherit property "getItem" with the proper type (2) 
PASS Storage interface: calling getItem(DOMString) on window.localStorage with too few arguments must throw TypeError 
PASS Storage interface: window.localStorage must inherit property "setItem" with the proper type (3) 
PASS Storage interface: calling setItem(DOMString,DOMString) on window.localStorage with too few arguments must throw TypeError 
PASS Storage interface: window.localStorage must inherit property "removeItem" with the proper type (4) 
PASS Storage interface: calling removeItem(DOMString) on window.localStorage with too few arguments must throw TypeError 
PASS Storage interface: window.localStorage must inherit property "clear" with the proper type (5) 
PASS StorageEvent interface: existence and properties of interface object 
PASS StorageEvent interface object length 
PASS StorageEvent interface object name 
FAIL StorageEvent interface: existence and properties of interface prototype object assert_equals: class string of StorageEvent.prototype expected "[object StorageEventPrototype]" but got "[object StorageEvent]"
PASS StorageEvent interface: existence and properties of interface prototype object's "constructor" property 
PASS StorageEvent interface: attribute key 
PASS StorageEvent interface: attribute oldValue 
PASS StorageEvent interface: attribute newValue 
PASS StorageEvent interface: attribute url 
PASS StorageEvent interface: attribute storageArea 
Harness: the test ran to completion.
